Confirm update available before removing current
The script now tries to download the chosen version of Go and only
proceeds to delete the currently installed version if the download was
successful.

+++ updateGo.sh
@@ -73,17 +73,22 @@ URL="https://go.dev/dl/go${VERSION}.${OS}-${ARCH}.tar.gz"
 # Check for updates to the script
 check_for_updates
 
-# Step 1: Remove the current version of Go if it exists
+# Step 1: Attempt to download the specified version of Go
+echo "Downloading Go version ${VERSION} from ${URL}..."
+wget $URL -O /tmp/go${VERSION}.${OS}-${ARCH}.tar.gz
+WGET_RESULT=$?
+if [ $WGET_RESULT != 0 ]; then
+  echo "Error: wget exited with code ${WGET_RESULT}"
+  exit 1
+fi
+
+# Step 2: Remove the current version of Go if it exists
 if [ -d "/usr/local/go" ]; then
   echo "Removing current Go installation..."
   sudo rm -rf /usr/local/go
 else
   echo "No existing Go installation found."
 fi
-
-# Step 2: Download the specified version of Go
-echo "Downloading Go version ${VERSION} from ${URL}..."
-wget $URL -O /tmp/go${VERSION}.${OS}-${ARCH}.tar.gz
 
 # Step 3: Extract the downloaded tarball to /usr/local
 echo "Extracting Go ${VERSION} to /usr/local..."
